qemu run.

+++ drivers/ide/ide-cd.h
@@ -434,7 +434,8 @@ struct atapi_mechstat_header {
 
 	byte     curlba[3];
 	byte     nslots;
-	__u8 short slot_tablelen;
+	//__u8 short slot_tablelen;
+	short slot_tablelen;
 };
